+++ WalletWasabi.Bridge/SoftwareWallet.cs
@@ -1,0 +1,10 @@
+using RawWallet = WalletWasabi.Wallets.Wallet;
+
+namespace WalletWasabi.Bridge;
+
+public class SoftwareWallet : ImprovedWallet
+{
+	public SoftwareWallet(RawWallet wallet) : base(wallet)
+	{
+	}
+}
